Move Zephyr project generation to new file
